Issue #11137: Code refactoring
A testbase to rule them all.
