+++ profile/modules/apps/os_publications/tests/src/ExistingSiteJavascript/RepecTest.php
@@ -3,14 +3,14 @@
 namespace Drupal\Tests\os_publications\ExistingSiteJavascript;
 
 use Drupal\repec\RepecInterface;
-use weitzman\DrupalTestTraits\ExistingSiteWebDriverTestBase;
+use Drupal\Tests\openscholar\ExistingSiteJavascript\TestBase;
 
 /**
  * RepecTest.
  *
  * @group functional-javascript
  */
-class RepecTest extends ExistingSiteWebDriverTestBase {
+class RepecTest extends TestBase {
 
   /**
    * Admin user.
@@ -61,43 +61,6 @@ class RepecTest extends ExistingSiteWebDriverTestBase {
     foreach ($template_fields as $field_label) {
       $web_assert->pageTextContains($field_label);
     }
-  }
-
-  /**
-   * Waits for the given time or until the given JS condition becomes TRUE.
-   *
-   * Shamelessly copied from DrupalCommerce.
-   *
-   * @param string $condition
-   *   JS condition to wait until it becomes TRUE.
-   * @param int $timeout
-   *   (Optional) Timeout in milliseconds, defaults to 1000.
-   * @param string $message
-   *   (optional) A message to display with the assertion. If left blank, a
-   *   default message will be displayed.
-   *
-   * @see \Behat\Mink\Driver\DriverInterface::evaluateScript()
-   *
-   * @throws \Behat\Mink\Exception\UnsupportedDriverActionException
-   * @throws \Behat\Mink\Exception\DriverException
-   */
-  protected function assertJsCondition($condition, $timeout = 1000, $message = '') {
-    $message = $message ?: "Javascript condition met:\n" . $condition;
-    $result = $this->getSession()->getDriver()->wait($timeout, $condition);
-    $this->assertNotEmpty($result, $message);
-  }
-
-  /**
-   * Waits for jQuery to become active and animations to complete.
-   *
-   * Shamelessly copied from DrupalCommerce.
-   *
-   * @throws \Behat\Mink\Exception\UnsupportedDriverActionException
-   * @throws \Behat\Mink\Exception\DriverException
-   */
-  protected function waitForAjaxToFinish() {
-    $condition = "(0 === jQuery.active && 0 === jQuery(':animated').length)";
-    $this->assertJsCondition($condition, 10000);
   }
 
 }
